Create triathlon award
Design a program that determines the award a person competing in a triathlon will receive.
Your program should read in the times (in minutes) for all three events of a triathlon, namely swimming, cycling, and running, and then calculate and display the total time taken to complete the triathlon.
The award a participant receives is based on the total time taken to complete the triathlon.  The qualifying time for awards is 100 minutes. Display the award that the participant will receive based on the following criteria.

Total time	                Award
Within qualifying time.	Provincial Colours
Within 5 minutes of qualifying time.	Provincial Half Colours
Within 10 minutes of qualifying time.	Provincial Scroll
More than 10 minutes of qualifying time.	No award