Create exception handling
( using filename = equations.txt)

Follow these steps:
Created a simple calculator application that asks a user to enter two numbers and the operation (e.g.+,-,x, etc.) that they’d like to perform on the numbers.  Display the answer to the equation.  Every equation entered by the user should be written to a text file.  Use defensive programming to write this program in a manner that is robust and handles unexpected events and user inputs.

Now extend your program to give the user the option to either enter two numbers and an operator, like before, or to read all of the equations from a new txt file (the user should add the name of the txt file as an input) and print out all of the equations together with the results.  Use defensive coding to ensure that the program does not crash if the file does not exist and the user is prompted again to enter the name of the file.